requirements.txt
absl-py==0.9.0
apptools==4.5.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==19.3.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
configobj==5.0.6
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
envisage==4.9.2
future==0.18.2
gast==0.3.3
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.8
imageio==2.8.0
importlib-metadata==1.6.1
jedi==0.17.2
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==2.6.11
MarkupSafe==1.1.1
matplotlib==3.2.2
mayavi==4.7.1
mistune==0.8.4
more-itertools==8.4.0
Mosek==9.2.12
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.18.5
open3d==0.10.0.0
opencv-python==4.2.0.34
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
py==1.8.2
pybind11==2.5.0
pycparser==2.20
pyface==7.0.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySide2==5.14.1
python-dateutil==2.8.1
pyzmq==19.0.2
requests==2.22.0
scipy==1.1.0
Send2Trash==1.5.0
shiboken2==5.14.1
six==1.15.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-hub==0.8.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
torch==1.5.0
torchfile==0.1.0
tornado==6.0.4
traitlets==4.3.3
traits==6.1.0
traitsui==7.0.0
urllib3==1.25.9
vtk==8.1.2
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
